# public/swagger.yaml
openapi: 3.0.3
info:
  title: Food Delivery System API
  description: |
    RESTful API for Food Delivery System with RAG Chatbot

    ## Features
    - User authentication & authorization
    - Restaurant & menu management
    - Order processing & tracking
    - Real-time delivery tracking
    - Push notifications
    - Chat messaging
    - RAG-powered chatbot

    ## Base URL
    - Development: `http://localhost:3000`
  version: 1.0.0
  contact:
    name: API Support
    email: support@fooddelivery.com
servers:
  - url: http://localhost:3000
    description: Development server
...
paths:
...
  /api/reviews:
    get:
      tags:
        - Reviews
      summary: Get reviews
      parameters:
        - name: restaurantId
          in: query
          schema:
            type: string
        - name: customerId
          in: query
          schema:
            type: string
        - name: limit
          in: query
          schema:
            type: integer
      responses:
        '200':
          description: List of reviews

    post:
      tags:
        - Reviews
      summary: Create review
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                orderId:
                  type: string
                customerId:
                  type: string
                restaurantId:
                  type: string
                foodRating:
                  type: integer
                  minimum: 1
                  maximum: 5
                deliveryRating:
                  type: integer
                  minimum: 1
                  maximum: 5
                overallRating:
                  type: integer
                  minimum: 1
                  maximum: 5
                comment:
                  type: string
      responses:
        '201':
          description: Review created
...